# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "product-api Documentation",
        "description": "Product API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "local server"
        }
    ],
    "paths": {
        "/api/product": {
            "get": {
                "summary": "Get a list of products",
                "operationId": "5cc57bec35d0f2045983cc8dcd00eccd",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/product/{id}": {
            "get": {
                "summary": "Get a product",
                "operationId": "2768d88590e39a7ded1d0032e64dded7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the product",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update a product",
                "operationId": "d3467c795490afbb49680821c9bd6b0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the product",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Product title",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Product description",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "description": {
                                        "description": "Product price",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "image of the product",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Product successfully updated"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "When an error occure"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}